29/ryancheley/requirements.txt
certifi==2019.11.28
chardet==3.0.4
confusable-homoglyphs==3.2.0
coreapi==2.3.3
coreschema==0.0.4
coverage==5.0.3
dj-database-url==0.5.0
Django==3.0.3
django-heroku==0.3.1
django-nose==1.4.6
django-registration==3.1
django-rest-swagger==2.2.0
djangorestframework==3.11.0
gunicorn==20.0.4
humanize==1.0.0
idna==2.8
itypes==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
nose==1.3.7
openapi-codec==1.3.2
psycopg2==2.8.4
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.0
tz==0.2.2
uritemplate==3.0.1
urllib3==1.25.8
whitenoise==5.0.1
